Refactor codes:
- create navigation name list in config.js
