+++ client/src/components/helper/useFetch/useFetch.js
@@ -1,6 +1,6 @@
 import { useState, useEffect } from 'react';
 import todoApi from '../../../Apis';
-import { INVALID_DATA, NETWORK_ERROR } from '../../../message/message';
+import { INVALID_DATA, NETWORK_ERROR } from '../../../config';
 
 const useFetch = setInitDataFunc => {
   const [isFetched, setIsFetched] = useState(false);
