--- client/src/message/message.js
@@ -1,4 +1,0 @@
-export const INVALID_DATA = '유효한 데이터가 없습니다.'
-export const NETWORK_ERROR = '서버에 접근할 수 없습니다.'
-export const LOADING = '안타깝게도 로딩중...';
-export const EMPTY_LIST = '빨리 할일 목록을 만들어내세요!'
